requirements.txt
uvicorn>=0.33.0
fastapi[all]>=0.115.0
orjson>=3.10.0
sqlalchemy>=2.0.0
asyncpg>=0.29.0
psycopg2-binary>=2.9.10
PyJWT==2.10.1
bcrypt